aeneas_pipeline/aeneas-func/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
azure-storage-blob
BeautifulSoup4>=4.5.1
lxml>=3.6.0
numpy>=1.9